small style changes and prettier
