+++ packages/nouns-webapp/src/components/VoteCard/index.tsx
@@ -10,8 +10,9 @@ import DelegateGroupedNounImageVoteTable from '../DelegateGroupedNounImageVoteTa
 import { useEthers } from '@usedapp/core';
 import responsiveUiUtilsClasses from '../../utils/ResponsiveUIUtils.module.css';
 import clsx from 'clsx';
-import { ensCacheKey, lookupAddress } from '../../utils/ensLookup';
+import { ensCacheKey } from '../../utils/ensLookup';
 import { useActiveLocale } from '../../hooks/useActivateLocale';
+import { lookupNNSOrENS } from '../../utils/lookupNNSOrENS';
 
 export enum VoteCardVariant {
   FOR,
@@ -77,7 +78,7 @@ const VoteCard: React.FC<VoteCardProps> = props => {
         return;
       }
 
-        lookupAddress(library, delegateInfo.delegate)
+      lookupNNSOrENS(library, delegateInfo.delegate)
         .then(name => {
           // Store data as mapping of address_Expiration => address or ENS
           if (name) {
