+++ packages/nouns-webapp/src/components/profileEvent/event/DesktopTransferEvent/index.tsx
@@ -34,7 +34,8 @@ const DesktopTransferEvent: React.FC<DesktopTransferEventProps> = props => {
             }}
           />
           <Trans>
-            Holder changed from <span
+            Holder changed from{' '}
+            <span
               data-tip={`View on Etherscan`}
               onClick={() => window.open(buildEtherscanAddressLink(event.from), '_blank')}
               data-for="view-on-etherscan-tooltip"
